requirements.txt
gunicorn
python-telegram-bot==20.7
python-dotenv
selenium
nest_asyncio
